==> 02-combine.qmd <==
---
title: "Combining the Datasets"
---

## Set up 

Loading the libraries 

```{r}
#| label: setup
#| message: false


library(tidyverse)
library(janitor)
library(lubridate)
```


## Importing VERSA 

```{r}
versa_data <- read_rds("data-processed/01-clean-versa.rds") |> glimpse()

```


## Import SPURS 

```{r}
spurs_data <- read_rds("data-processed/01-clean-spurs.rds") |> glimpse()
```


## Combine both clean data sets

```{r}
data_combine <- versa_data |> 
  bind_rows(spurs_data)

data_combine
```

## Export combined file 

```{r}
data_combine |>
write_rds("data-processed/02-combine.rds")
```
